libshvcoreqt: Put headers directly into include/
