+++ libshvcoreqt/include/shv/coreqt/shvcoreqtglobal.h
